Made previous fix complete
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@39159 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/app/oligofar/cseqscanner.hpp
@@ -116,7 +116,7 @@ protected:
     class C_LoopImpl_ColorspAmbiguities : public C_LoopImpl_Ncbi8naAmbiguities
     {
     public:
-        C_LoopImpl_ColorspAmbiguities( int windowLength, double maxSimpl, int maxAlt, unsigned mask ) : 
+        C_LoopImpl_ColorspAmbiguities( int windowLength, double maxSimpl, Uint8 maxAlt, unsigned mask ) : 
             C_LoopImpl_Ncbi8naAmbiguities( windowLength, maxSimpl, maxAlt, mask ) {}
         void Prepare( char a );
         void Update( char a );
